backend/requirements.txt

# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic

# PostgreSQL dependencies
sqlalchemy==2.0.41
greenlet==3.2.4
psycopg2-binary
asyncpg

# Redis dependencies
redis[hiredis]==5.0.1

# Additional utilities
python-multipart
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
